+++ src/vellum/lib/__init__.py
@@ -1,4 +1,4 @@
-from .test_suite import VellumTestSuite
+from .test_suites import VellumTestSuite
 
 __all__ = [
     "VellumTestSuite",
