+++ src/vellum/lib/test_suites/__init__.py
@@ -1,0 +1,5 @@
+from .resources import VellumTestSuite
+
+__all__ = [
+    "VellumTestSuite",
+]
