+++ src/vellum/lib/test_suites/constants.py
@@ -1,0 +1,2 @@
+DEFAULT_POLLING_INTERVAL_MS = 500
+DEFAULT_MAX_POLLING_DURATION_MS = 30 * 60 * 1000
