+++ src/vellum/lib/test_suites/exceptions.py
@@ -1,0 +1,2 @@
+class TestSuiteRunResultsException(Exception):
+    pass
